fix(eng-analytics): rename test_health logic module to suite_health
